backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
kafka-python==2.0.2
minio==7.2.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.2
joblib==1.3.2
python-multipart==0.0.6

